+++ src/index.ts
@@ -1,3 +1,4 @@
 export { MetaConfig } from './models/meta-config';
 export { MetaService } from './meta.service';
 export { MetaModule } from './meta.module';
+export { MetaGuard } from './meta.guard';