+++ src/meta.module.ts
@@ -2,7 +2,9 @@ import { NgModule, ModuleWithProviders } from '@angular/core';
 import { Optional, SkipSelf } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import { MetaService } from './meta.service';
-import { MetaConfig, META_CONFIG_TOKEN } from './models/meta-config';
+import { DEFAULT_META_CONFIG, MetaConfig } from './models/meta-config';
+import { META_CONFIG_TOKEN } from './models/meta-constants';
+import { MetaGuard } from './meta.guard';
 
 @NgModule({
   imports: [
@@ -10,11 +12,12 @@ import { MetaConfig, META_CONFIG_TOKEN } from './models/meta-config';
   ]
 })
 export class MetaModule {
-  public static forRoot(metaConfig: MetaConfig = { useTitleSuffix: false, defaults: {} }): ModuleWithProviders {
+  public static forRoot(metaConfig: MetaConfig = DEFAULT_META_CONFIG): ModuleWithProviders {
     return {
       ngModule: MetaModule,
       providers: [
         { provide: META_CONFIG_TOKEN, useValue: metaConfig },
+        MetaGuard,
         MetaService
       ]
     };